Update readme.md
At the moment, pressing on "List Comprehension" in readme.md, navigates you to a "404" response page. The reason for this, is because it is currently navigating to "../13_Day_List_comprehension/13_list_comprehension.md", going up one directory too much. As such, it should be changed to "./13_Day_List_comprehension/13_list_comprehension.md" (one less period in the beginning)

Please let me know of any concerns!