Update 02_variables_builtin_functions.md
You can't use int() on a string with the value '10.6' directly because int() is meant for converting strings that represent whole numbers. When you try to convert a string representing a decimal number to an integer, it results in a ValueError. To handle this, you first need to convert the string to a float using float(), and then you can convert the float to an integer using int().